==> course completion/data_construction/11_trend_predictors.ipynb <==
import pickle
import glob
import pandas as pd

enrl = glob.glob("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\term_enrl_intensity_trend_values\*.p")
enrl_dict = {}
for f in enrl:
    d = pickle.load(open(f, "rb"))
    enrl_dict = {**enrl_dict, **d}

gpa = glob.glob("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\term_gpa_trend_values\*.p")
gpa_dict = {}
for f in gpa:
    d = pickle.load(open(f, "rb"))
    gpa_dict = {**gpa_dict, **d}

enrl_df = pd.DataFrame.from_dict(enrl_dict, orient='index').reset_index()
enrl_df.loc[:,'vccsid'] = enrl_df['index'].apply(lambda x: x[0])
enrl_df.loc[:,'strm'] = enrl_df['index'].apply(lambda x: x[-1])
enrl_df = enrl_df.drop(['index'], axis=1)
enrl_df = enrl_df.rename(columns={0:'enrl_intensity_trend'})

gpa_df = pd.DataFrame.from_dict(gpa_dict, orient='index').reset_index()
gpa_df.loc[:,'vccsid'] = gpa_df['index'].apply(lambda x: x[0])
gpa_df.loc[:,'strm'] = gpa_df['index'].apply(lambda x: x[-1])
gpa_df = gpa_df.drop(['index'], axis=1)
gpa_df = gpa_df.rename(columns={0:'gpa_trend'})

trend_df = enrl_df.merge(gpa_df, how='inner', on=['vccsid', 'strm']).loc[:,['vccsid','strm','gpa_trend','enrl_intensity_trend']]

all_vccsid_strm = pd.read_stata("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\all_vccsid_strm.dta").loc[:,['vccsid', 'strm']]

trend_df = all_vccsid_strm.merge(trend_df, how='left', on=['vccsid', 'strm']).sort_values(['vccsid','strm']).loc[:,['vccsid','strm','gpa_trend','enrl_intensity_trend']]

trend_df.to_stata("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\trend_predictors.dta", write_index=False)
