==> course completion/data_construction/24_finalize_term_specific_predictors.ipynb <==
import pandas as pd
import numpy as np

all_vccsid_strm_1 = pd.read_stata("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\all_observations.dta")\
.loc[:,['vccsid','strm']].drop_duplicates()
all_vccsid_strm_2 = pd.read_stata("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\all_vccsid_strm.dta")
all_vccsid_strm = all_vccsid_strm_1.merge(all_vccsid_strm_2, on=['vccsid', 'strm'], how="inner")
all_vccsid_strm.shape

sum(all_vccsid_strm.first_strm == all_vccsid_strm.strm), sum(all_vccsid_strm.first_strm == all_vccsid_strm.strm)/all_vccsid_strm.shape[0]

all_vccsid_strm.loc[:,'first_ind'] = all_vccsid_strm.apply(lambda x: x.strm == x.first_strm, axis=1).astype(int)
all_vccsid_strm = all_vccsid_strm.drop(['first_strm'], axis=1)

from collections import Counter
Counter(all_vccsid_strm.first_ind)

d1 = pd.read_stata('C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\age_gpa_earn.dta')
d2 = pd.read_stata('C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\four_additional_predictors.dta')
d3 = pd.read_stata('C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\twelve_additional_predictors.dta')
d4 = pd.read_stata('C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\trend_predictors.dta')
d5 = pd.read_stata('C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\ever_dually_enrolled.dta')

merged = all_vccsid_strm.merge(d1, how='inner', on=['vccsid','strm']).\
merge(d2, how='inner', on=['vccsid','strm']).\
merge(d3, how='inner', on=['vccsid','strm']).\
merge(d4, how='inner', on=['vccsid','strm']).\
merge(d5, how='inner', on=['vccsid','strm'])

merged.shape

merged_part1 = merged[merged.first_ind == 1]
merged_part2 = merged[merged.first_ind == 0]

merged_part1 = merged_part1.fillna(0)

pd.isnull(merged_part2).any()

impute_list_1 = set(["cum_gpa", "overall_prop_comp", "prop_comp_sd"])

def impute(train, test):
    for p in impute_list_1:
        avg_p = np.nanmean(train[p])
        train.loc[:,p] = train.loc[:,p].apply(lambda x: avg_p if pd.isnull(x) else x)
        test.loc[:,p] = test.loc[:,p].apply(lambda x: avg_p if pd.isnull(x) else x)
    return train, test               

train_df, test_df = impute(merged_part2[merged_part2.strm < 2192], merged_part2[merged_part2.strm >= 2192])
merged_part2_new = pd.concat([train_df, test_df])

merged_new = pd.concat([merged_part1, merged_part2_new])

merged_new.to_csv("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\term_specific_predictors.csv", index=False)

assert pd.isnull(merged_new).any().any() == False
merged_new.shape
