==> course completion/data_construction/25_create_cluster_specific_predictors.ipynb <==
import pandas as pd
import numpy as np

all_vccsid_strm_1 = pd.read_stata("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\all_observations.dta")
all_vccsid_strm_2 = pd.read_stata("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\all_vccsid_strm.dta")
df1 = all_vccsid_strm_1.merge(all_vccsid_strm_2, on=['vccsid', 'strm'], how="inner")
df1.shape

df2 = pd.read_stata("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\prior_courses.dta")

df1 = df1[df1.course.apply(lambda x: len(x) == 7)]
df1 = df1[df1.grade.apply(lambda x: x in {"A","B","C","D","F","W"})]
df1.shape

all_subjects = []
for cn in list(df1.course) + list(df2.course):
    all_subjects.append(cn.split("_")[0])
all_subjects = np.unique(all_subjects)

all_subjects

subj_dict = {'BUS': ['ACC', 'BUS', 'MKT', 'HRI', 'FIN', 'REA', 'ACQ', 'CON', 'ISR'],
             'ART': ['ART', 'DAN', 'MUS', 'HRT', 'CRF', 'PHT', 'IDS', "DEC"],
             'SCI': ['BIO', 'GOL', 'PHY', 'NAS', 'CHM', 'SCT'],
             'EGR': ['EGR', 'CSC', 'ITD', 'ITE', 'ITN', 'ITP',
                     'HIM', 'ENE', 'GIS', 'ENV', 'HIT', 'ARC',
                     'ETR', 'AST', 'IND', 'IST', 'DRF', 'CAD', 
                     'TEL', 'MEC', 'AMT', 'ARO', 'CIV', "INS", 'NUC', 'UMS', "AVI", "ESR", "NAN", "SIM"],
             'MTH': ['MTE', 'MTH', 'MTT', 'MCR', 'MDE'],
             'SOC': ['EDU', 'SOC', 'ECO', 'GEO', 'HIS', 'PSY',
                     'PLS', 'ADJ', 'LGL', 'CHD', 'SSC', "PBS", "EIP"],
             'MED': ['EMS', 'NUR', 'HLT', 'PED', 'DMS', 'RAD', 
                     'EMT', 'PNE', 'DNA', 'SUR', 'NSG', 'HCT',
                     'DNH', 'OCT', 'CSP', 'DIT', 'PSG', 'PNG',
                     'PTH', 'MEN', 'MDL', 'RTH', "MDA", "OPT", "ROC", "DNL", "OMP", "PBH", 
                     'DIA'],
             'HUM': ['HUM', 'CST', 'PHI', 'REL', 'ENG', 'ENF',
                     'EDE', 'HMS', 'SDV', 'SPD', 'STD', 'SCM', 'MET'],
             'FLA': ['ASL', 'ARA', 'FRE', 'GER', 'JPN', 'KOR',
                     'ITA', 'LAT', 'SPA', 'CHI', 'ESL', 'RUS', 'INT', 'GRE', 'POR', 'VTN', "HIN"],
             'OCC': ['FST', 'AGR', 'AIR', 'ARO', 'AUT', 'FOR',
                     'FIR', 'WEL', 'VET', 'EQU', 'LBR', 'BCS',
                     'BLD', 'BSK', 'COS', 'VEN', 'TRK', 'TRV',
                     'PNT', 'MSC', 'MAR', 'MAC', 'ELE', 'RVH',
                     'RPK', 'SAF', 'AUB', "DSL", 'FNS', "MRT",
                     'MTS', 'RPK', "FUR", "GWR", "MIN", "PPT",
                     'HVE', 'APL']}
subj_inverse_dict = {e:k for k,v in subj_dict.items() for e in v}
assert len(subj_inverse_dict) >= len(all_subjects)

df2.loc[:,'subject'] = df2.course.apply(lambda x: x.split("_")[0])
df2.loc[:,'cluster'] = df2.subject.apply(lambda x: subj_inverse_dict[x])
df2.loc[:,'grade_point'] = df2.credit * df2.est_grade

cluster_dict = {}
for strm in [2142, 2143, 2144, 2152, 2153, 2154, 2162, 2163, 2164, 2172, 2173, 2174, 2182, 2183, 2184, 2192, 2193, 2194]:
    df2_sub = df2[df2.strm < strm].copy()
    if df2_sub.shape[0] > 0:
        df2_sub_agg = df2_sub.groupby(['vccsid', 'cluster']).agg({'credit':'sum', 'grade_point':'sum'}).reset_index()
        df2_sub_agg.loc[:,'est_grade'] = df2_sub_agg.grade_point/df2_sub_agg.credit
        for i in range(df2_sub_agg.shape[0]):
            vccsid = df2_sub_agg.vccsid.iloc[i]
            cluster = df2_sub_agg.cluster.iloc[i]
            est_grade = df2_sub_agg.est_grade.iloc[i]
            try:
                cluster_dict[vccsid + "_" + str(strm)].append((cluster, est_grade))
            except KeyError:
                cluster_dict[vccsid + "_" + str(strm)] = [(cluster, est_grade)]

df1.loc[:,'subject'] = df1.course.apply(lambda x: x.split("_")[0])
df1.loc[:,'cluster'] = df1.subject.apply(lambda x: subj_inverse_dict[x])

full_cluster_dict = {}
for i in range(df1.shape[0]):
    vccsid = df1.vccsid.iloc[i]
    strm = df1.strm.iloc[i]
    college = df1.college.iloc[i]
    course = df1.course.iloc[i]
    section  = df1.section.iloc[i]
    subject = df1.subject.iloc[i]
    cluster = df1.cluster.iloc[i]
    l = []
    if vccsid + "_" + str(strm) in cluster_dict:
        ll = cluster_dict[vccsid + "_" + str(strm)]
        l = [(t[0]+"_"+cluster, t[1]) for t in ll]
    d = {}
    for t in l:
        d[t[0]] = 1
        d[t[0] + "_grade"] = t[1]
    full_cluster_dict["-".join([vccsid, str(strm), college, course, section])] = d.copy()

full_cluster_df = pd.DataFrame.from_dict(full_cluster_dict, orient='index')
for indx,var in enumerate(['vccsid','strm','college','course','section']):
    full_cluster_df.loc[:,var] = list(pd.Series(list(full_cluster_df.index)).apply(lambda x: x.split("-")[indx]))
full_cluster_df.loc[:,'strm'] = full_cluster_df.strm.astype(int)
full_cluster_df = df1.loc[:,['vccsid','strm','college','course','section']].merge(full_cluster_df, how='left',
                                                                                  on = ['vccsid','strm','college','course','section']).fillna(0)

full_cluster_df.to_csv("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\cluster_specific_predictors.csv", index=False)

full_cluster_df.shape
