==> course completion/data_construction/26_create_concurrent_predictors.ipynb <==
import pandas as pd
import numpy as np

concurrent_courses = pd.read_stata("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\concurrent_courses.dta")

avg_grades = pd.read_stata("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\avg_grades.dta")
avg_dict = {}
avg_dict_2 = {}
for j in range(avg_grades.shape[0]):
    course = avg_grades.course.iloc[j]
    strm = avg_grades.strm.iloc[j]
    college = avg_grades.college.iloc[j]
    avg_dict[(course,strm,college)] = avg_grades.avg_grade.iloc[j]
    avg_dict_2[(strm,college)] = avg_grades.avg_grade_2.iloc[j]

avg_g = []
for i in range(concurrent_courses.shape[0]):
    course = concurrent_courses.course.iloc[i]
    strm = concurrent_courses.strm.iloc[i]
    college = concurrent_courses.college.iloc[i]
    try:
        avg_g.append(avg_dict[(course,strm,college)])
    except KeyError:
        avg_g.append(avg_dict_2[(strm,college)])
concurrent_courses.loc[:,'avg_g'] = avg_g

vccsid_sets = []
l = list(np.unique(concurrent_courses.vccsid))
n = len(l) // 2000 + 1
for i in range(2000):
    vccsid_sets.append(l[(n*i):(n*(i+1))])

vccsid_sets = [l for l in vccsid_sets if len(l) > 0]
len(vccsid_sets)

all_parts = []
for i,ll in enumerate(vccsid_sets):
    print(i)
    ll_series = pd.DataFrame({'vccsid':ll})
    part = ll_series.merge(concurrent_courses, how='inner', on=['vccsid'])
    print(part.shape[0])
    all_parts.append(part.copy())
    print("")

def compute_concurrent(t):
    new_t = t.groupby(['vccsid','strm','college','course']).agg('sum').reset_index()
    new_t.loc[:,'avg_g_concurrent'] = new_t.avg_g/new_t.credit
    new_t = new_t.drop(['avg_g', 'credit'], axis=1)
    return new_t

part_new_list = []
for a,part in enumerate(all_parts):
    print(a)
    part.loc[:,'avg_g'] = part.avg_g * part.credit
    all_vccsid_strm = part.loc[:,['vccsid','strm']].drop_duplicates()
    n = all_vccsid_strm.shape[0]
    temp_list = []
    for i in range(n):
        part_sub = all_vccsid_strm.iloc[i:(i+1),:].merge(part, on=['vccsid','strm'])
        indx = []
        indx_2 = []
        m = part_sub.shape[0]
        for j in range(m):
            indx.extend([k for k in range(m) if k != j])
            indx_2.extend([j]*(m-1))
        part_sub1 = part_sub.iloc[indx,:].loc[:,['credit', 'avg_g']]
        part_sub1.index = range(m*(m-1))
        part_sub2 = part_sub.iloc[indx_2,:].loc[:,['vccsid','strm','college', 'course']]
        part_sub2.index = range(m*(m-1))
        temp = pd.concat([part_sub2, part_sub1], axis=1)
        temp_list.append(temp)
    part_new = pd.concat(temp_list)
    part_new = compute_concurrent(part_new)
    part_new_list.append(part_new.copy())

merged = pd.concat(part_new_list)

merged.describe()

target = pd.read_csv("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\course_specific_predictors.csv")

target = target.merge(merged, on = ['vccsid', 'strm', 'college', 'course'], how='left')

target.loc[:,'has_avg_g_concurrent'] = ~pd.isnull(target.avg_g_concurrent)

target.loc[:,'has_avg_g_concurrent'] = target.loc[:,'has_avg_g_concurrent'].astype(int)

target = target.fillna(0)

target.to_csv("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\course_specific_predictors_new.csv", index=False)
target.shape

term = pd.read_csv("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\term_specific_predictors.csv")
term.shape

new1 = pd.read_stata("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\additional_term_specific.dta")
new2 = pd.read_stata("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\cip_degree.dta")

term = term.merge(new1, on=['vccsid', 'strm'], how='inner').merge(new2, on=['vccsid','strm'], how='inner')
term.shape

term.to_csv("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\term_specific_predictors_new.csv", index=False)
