==> course completion/data_construction/27_create_course_specific_predictors.ipynb <==
import pickle
import pandas as pd
import numpy as np

vccs_prerequisite_dict = pickle.load(open("C:/Users/ys8mz/Box Sync/GAA Transfer/data/intermediate_files/ys8mz/degree_audit_analyses/vccs_prerequisite_dict_2021.p", "rb"))
# vccs_prerequisite_dict['MTH_263'] = [['MTH_167'], ['MTH_161', 'MTH_162'], ['MTH_163', 'MTH_164']]
# vccs_prerequisite_dict['EGR_126'] = [['CSC_110']]

all_vccsid_strm_1 = pd.read_stata("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\all_observations.dta").loc[:,['vccsid','college','strm','course','section']].drop_duplicates()
all_vccsid_strm_2 = pd.read_stata("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\all_vccsid_strm.dta")
all_observations = all_vccsid_strm_1.merge(all_vccsid_strm_2, on=['vccsid', 'strm'], how="inner")
all_observations.shape

all_observations.loc[:,['vccsid', 'college', 'strm', 'course']].drop_duplicates().shape

prior_courses = pd.read_stata("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\prior_courses.dta")
prior_courses = prior_courses.merge(all_observations.loc[:,['vccsid']].drop_duplicates(), how='inner', on=['vccsid'])
print(prior_courses.shape)
target_courses = pd.read_stata("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\part1_new.dta")
target_courses = target_courses.merge(all_observations, how='inner', on=['vccsid','college','strm','course','section'])
print(target_courses.shape)

def create_cdict(df):
    cdict = {}
    for j in range(df.shape[0]):
        course = df.course.iloc[j]
        credit = df.credit.iloc[j]
        est_grade = df.est_grade.iloc[j]
        try:
            cdict[course].append((credit,est_grade))
        except KeyError:
            cdict[course] = [(credit,est_grade)]
    return cdict.copy()

def calc_avg_grade(t):
    tot_cred = 0
    tot_grade_point = 0
    for p in t:
        tot_cred += p[0]
        tot_grade_point += p[0]*p[1]
    return tot_grade_point/tot_cred

vccsid_sets = []
l = list(np.unique(target_courses.vccsid))
n = len(l) // 500 + 1
for i in range(500):
    vccsid_sets.append(l[(n*i):(n*(i+1))])

for i,ll in enumerate(vccsid_sets):
    print(i)
    ll_series = pd.DataFrame({'vccsid':ll})
    prior_part = ll_series.merge(prior_courses, how='inner', on=['vccsid'])
    target_part = ll_series.merge(target_courses, how='inner', on=['vccsid'])
    print(prior_part.shape[0], target_part.shape[0])
    prior_part.to_stata("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\prior_courses\\part{}.dta".format(i),
                        write_index=False)
    target_part.to_stata("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\target_courses\\part{}.dta".format(i),
                        write_index=False)
    print("")

# has_prereq_grade = []
# prereq_grade = []
# for i in range(target_courses.shape[0]):
#     if i % 1000 == 0:
#         print(i)
#     vccsid = target_courses.vccsid.iloc[i]
#     strm = target_courses.strm.iloc[i]
#     course = target_courses.course.iloc[i]
#     prior_sub = prior_courses[prior_courses.vccsid == vccsid]
#     prior_sub = prior_sub[prior_sub.strm < strm]
#     if prior_sub.shape[0] == 0:
#         has_prereq_grade.append(0)
#         prereq_grade.append(0)
#     else:
#         if course not in vccs_prerequisite_dict:
#             has_prereq_grade.append(0)
#             prereq_grade.append(0)
#             continue
#         else:
#             prerequisite_l = []
#             for l in vccs_prerequisite_dict[course]:
#                 prerequisite_l.extend(l)
#             cdict = create_cdict(prior_sub)
#             clist = list(cdict.keys())
#             common_l = np.intersect1d(prerequisite_l, clist)
#             if len(common_l) == 0:
#                 has_prereq_grade.append(0)
#                 prereq_grade.append(0)
#             else:
#                 new_t = []
#                 for c in common_l:
#                     new_t.extend(cdict[c])
#                 has_prereq_grade.append(1)
#                 prereq_grade.append(calc_avg_grade(new_t))

prereq_r = []
for i in range(500):
    prereq_r.extend(pickle.load(open("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\prereq\\part{}.p".format(i), "rb")))

prereq_df = pd.DataFrame(prereq_r, columns = ['vccsid', 'strm', 'college', 'course', 'section', 'has_prereq_grade', 'prereq_grade'])

print(target_courses.shape[0])
target_courses = target_courses.merge(prereq_df, how='inner', on=['vccsid', 'strm', 'college', 'course', 'section'])
print(target_courses.shape[0])

# has_repeat_grade = []
# repeat_grade = []
# for i in range(target_courses.shape[0]):
#     if i % 1000 == 0:
#         print(i)
#     vccsid = target_courses.vccsid.iloc[i]
#     strm = target_courses.strm.iloc[i]
#     course = target_courses.course.iloc[i]
#     prior_sub = prior_courses[prior_courses.vccsid == vccsid]
#     prior_sub = prior_sub[prior_sub.strm < strm]
#     if prior_sub.shape[0] == 0:
#         has_repeat_grade.append(0)
#         repeat_grade.append(0)
#     else:
#         cdict = create_cdict(prior_sub)
#         if course not in cdict:
#             has_repeat_grade.append(0)
#             repeat_grade.append(0)
#         else:
#             has_repeat_grade.append(1)
#             repeat_grade.append(calc_avg_grade(cdict[course]))

repeat_r = []
for i in range(500):
    repeat_r.extend(pickle.load(open("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\repeat\\part{}.p".format(i), "rb")))

repeat_df = pd.DataFrame(repeat_r, columns = ['vccsid', 'strm', 'college', 'course', 'section', 'has_repeat_grade', 'repeat_grade'])

print(target_courses.shape[0])
target_courses = target_courses.merge(repeat_df, how='inner', on=['vccsid', 'strm', 'college', 'course', 'section'])
print(target_courses.shape[0])

# target_courses.loc[:,'has_repeat_grade'] = has_repeat_grade
# target_courses.loc[:,'repeat_grade'] = repeat_grade
# target_courses.loc[:,'has_prereq_grade'] = has_prereq_grade
# target_courses.loc[:,'prereq_grade'] = prereq_grade

avg_grades = pd.read_stata("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\avg_grades.dta")
avg_dict = {}
avg_dict_2 = {}
for j in range(avg_grades.shape[0]):
    course = avg_grades.course.iloc[j]
    strm = avg_grades.strm.iloc[j]
    college = avg_grades.college.iloc[j]
    avg_dict[(course,strm,college)] = avg_grades.avg_grade.iloc[j]
    avg_dict_2[(strm,college)] = avg_grades.avg_grade_2.iloc[j]

avg_g = []
for i in range(target_courses.shape[0]):
    course = target_courses.course.iloc[i]
    strm = target_courses.strm.iloc[i]
    college = target_courses.college.iloc[i]
    try:
        avg_g.append(avg_dict[(course,strm,college)])
    except KeyError:
        avg_g.append(avg_dict_2[(strm,college)])
target_courses.loc[:,'avg_g'] = avg_g

target_courses.loc[:,'summer_ind'] = target_courses.strm.apply(lambda x: int(x % 10 == 3))

target_courses.columns.values

target_courses = target_courses.loc[:,['vccsid','strm','college','course','section',
                                       'summer_ind', 'section_size', 'eve_ind', 'online_ind', 'lvl2_ind',
                                       'has_repeat_grade', 'repeat_grade', 'has_prereq_grade', 'prereq_grade', 
                                       'avg_g']]
print(target_courses.shape)

target_courses.to_csv("C:\\Users\\ys8mz\\Box Sync\\Clickstream\\data\\bias\\course_specific_predictors.csv", index=False)
